requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1
python-dotenv==1.0.1
matplotlib==3.8.3